Split mutex and barriers in pal into separate header
