+++ include/ccf/endpoint_registry.h
@@ -5,7 +5,7 @@
 #include "ccf/ds/json_schema.h"
 #include "ccf/endpoint.h"
 #include "ccf/endpoint_context.h"
-#include "ccf/pal.h"
+#include "ccf/pal/locking.h"
 #include "ccf/rpc_context.h"
 #include "ccf/tx.h"
