+++ include/ccf/pal.h
@@ -18,7 +18,6 @@
 #  include "crypto/ecdsa.h"
 
 #  include <cstring>
-#  include <mutex>
 #else
 #  include "ccf/ds/ccf_exception.h"
 #  include "ccf/pal/attestation_sgx.h"
@@ -27,7 +26,6 @@
 #  include <openenclave/attestation/attester.h>
 #  include <openenclave/bits/defs.h>
 #  include <openenclave/bits/security.h>
-#  include <openenclave/edger8r/enclave.h> // For oe_lfence
 #  include <openenclave/enclave.h>
 #  include <openenclave/log.h>
 #  include <openenclave/tracee.h>
@@ -58,19 +56,10 @@ namespace ccf::pal
 
 #if !defined(INSIDE_ENCLAVE) || defined(VIRTUAL_ENCLAVE)
 
-  /**
-   * Virtual enclaves and the host code share the same PAL.
-   * This PAL takes no dependence on OpenEnclave, but also does not apply
-   * security hardening.
-   */
-  using Mutex = std::mutex;
-
   static inline void* safe_memcpy(void* dest, const void* src, size_t count)
   {
     return ::memcpy(dest, src, count);
   }
-
-  static inline void speculation_barrier() {}
 
   static inline void redirect_platform_logging() {}
 
@@ -291,52 +280,9 @@ namespace ccf::pal
   }
 
 #else
-  /**
-   * Temporary workaround until the fix for
-   * https://github.com/openenclave/openenclave/issues/4555 is available in a
-   * release.
-   */
-  class MutexImpl
-  {
-  private:
-    pthread_spinlock_t sl;
-
-  public:
-    MutexImpl()
-    {
-      pthread_spin_init(&sl, PTHREAD_PROCESS_PRIVATE);
-    }
-
-    ~MutexImpl()
-    {
-      pthread_spin_destroy(&sl);
-    }
-
-    void lock()
-    {
-      pthread_spin_lock(&sl);
-    }
-
-    bool try_lock()
-    {
-      return pthread_spin_trylock(&sl) == 0;
-    }
-
-    void unlock()
-    {
-      pthread_spin_unlock(&sl);
-    }
-  };
-
-  using Mutex = MutexImpl;
   static inline void* safe_memcpy(void* dest, const void* src, size_t count)
   {
     return oe_memcpy_with_barrier(dest, src, count);
-  }
-
-  static inline void speculation_barrier()
-  {
-    oe_lfence();
   }
 
   static void open_enclave_logging_callback(
