+++ src/ds/test/stub_writer.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/pal.h"
+#include "ccf/pal/locking.h"
 #include "ds/ring_buffer_types.h"
 
 #include <limits>
