+++ src/kv/store.h
@@ -5,6 +5,7 @@
 #include "apply_changes.h"
 #include "ccf/ds/ccf_exception.h"
 #include "ccf/kv/read_only_store.h"
+#include "ccf/pal/locking.h"
 #include "consensus/aft/request.h"
 #include "deserialise.h"
 #include "kv/committable_tx.h"
