+++ src/node/acme_client.h
@@ -11,6 +11,7 @@
 #include "ccf/ds/logger.h"
 #include "ccf/http_consts.h"
 #include "ccf/http_status.h"
+#include "ccf/pal/locking.h"
 #include "ds/messaging.h"
 #include "ds/thread_messaging.h"
 #include "http/http_parser.h"
