+++ src/node/node_to_node_channel_manager.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/pal.h"
+#include "ccf/pal/locking.h"
 #include "channels.h"
 #include "node/node_to_node.h"
