+++ src/node/test/historical_queries.cpp
@@ -8,6 +8,7 @@
 #include "node/historical_queries.h"
 
 #include "ccf/crypto/rsa_key_pair.h"
+#include "ccf/pal/locking.h"
 #include "ds/messaging.h"
 #include "ds/test/stub_writer.h"
 #include "kv/test/null_encryptor.h"
